refactor chain service
