+++ beacon-chain/blockchain/service.go
@@ -8,22 +8,13 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
-	gethTypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/gogo/protobuf/proto"
 	"github.com/prysmaticlabs/prysm/beacon-chain/attestation"
 	b "github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/state"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
 	"github.com/prysmaticlabs/prysm/beacon-chain/powchain"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
-	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/event"
 	"github.com/prysmaticlabs/prysm/shared/hashutil"
-	handler "github.com/prysmaticlabs/prysm/shared/messagehandler"
-	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/sirupsen/logrus"
 )
 
@@ -104,7 +95,6 @@ func (c *ChainService) Start() {
 		log.Info("Beacon chain data already exists, starting service")
 		c.genesisTime = time.Unix(int64(beaconState.GenesisTime), 0)
 		c.finalizedEpoch = beaconState.FinalizedEpoch
-		go c.blockProcessing()
 	} else {
 		log.Info("Waiting for ChainStart log from the Validator Deposit Contract to start the beacon chain...")
 		if c.web3Service == nil {
@@ -114,29 +104,34 @@ func (c *ChainService) Start() {
 		subChainStart := c.web3Service.ChainStartFeed().Subscribe(c.chainStartChan)
 		go func() {
 			genesisTime := <-c.chainStartChan
-			initialDepositsData := c.web3Service.ChainStartDeposits()
-			initialDeposits := make([]*pb.Deposit, len(initialDepositsData))
-			for i := range initialDepositsData {
-				initialDeposits[i] = &pb.Deposit{DepositData: initialDepositsData[i]}
-			}
-
-			depositRoot := c.web3Service.DepositRoot()
-			latestBlockHash := c.web3Service.LatestBlockHash()
-			eth1Data := &pb.Eth1Data{
-				DepositRootHash32: depositRoot[:],
-				BlockHash32:       latestBlockHash[:],
-			}
-			beaconState, err := c.initializeBeaconChain(genesisTime, initialDeposits, eth1Data)
-			if err != nil {
-				log.Fatalf("Could not initialize beacon chain: %v", err)
-			}
-			c.finalizedEpoch = beaconState.FinalizedEpoch
-			c.stateInitializedFeed.Send(genesisTime)
-			c.canonicalStateFeed.Send(beaconState)
-			go c.blockProcessing()
-			subChainStart.Unsubscribe()
+			c.processChainStartTime(genesisTime, subChainStart)
 		}()
 	}
+}
+
+// processChainStartTime initializes a series of deposits from the ChainStart deposits in the eth1
+// deposit contract, initializes the beacon chain's state, and kicks off the beacon chain.
+func (c *ChainService) processChainStartTime(genesisTime time.Time, chainStartSub event.Subscription) {
+	initialDepositsData := c.web3Service.ChainStartDeposits()
+	initialDeposits := make([]*pb.Deposit, len(initialDepositsData))
+	for i := range initialDepositsData {
+		initialDeposits[i] = &pb.Deposit{DepositData: initialDepositsData[i]}
+	}
+
+	depositRoot := c.web3Service.DepositRoot()
+	latestBlockHash := c.web3Service.LatestBlockHash()
+	eth1Data := &pb.Eth1Data{
+		DepositRootHash32: depositRoot[:],
+		BlockHash32:       latestBlockHash[:],
+	}
+	beaconState, err := c.initializeBeaconChain(genesisTime, initialDeposits, eth1Data)
+	if err != nil {
+		log.Fatalf("Could not initialize beacon chain: %v", err)
+	}
+	c.finalizedEpoch = beaconState.FinalizedEpoch
+	c.stateInitializedFeed.Send(genesisTime)
+	c.canonicalStateFeed.Send(beaconState)
+	chainStartSub.Unsubscribe()
 }
 
 // initializes the state and genesis block of the beacon chain to persistent storage
@@ -230,370 +225,4 @@ func (c *ChainService) doesPoWBlockExist(hash [32]byte) bool {
 	}
 
 	return powBlock != nil
-}
-
-// blockProcessing subscribes to incoming blocks, processes them if possible, and then applies
-// the fork-choice rule to update the beacon chain's head.
-func (c *ChainService) blockProcessing() {
-	subBlock := c.incomingBlockFeed.Subscribe(c.incomingBlockChan)
-	defer subBlock.Unsubscribe()
-	for {
-		select {
-		case <-c.ctx.Done():
-			log.Debug("Chain service context closed, exiting goroutine")
-			return
-
-		// Listen for a newly received incoming block from the feed. Blocks
-		// can be received either from the sync service, the RPC service,
-		// or via p2p.
-		case block := <-c.incomingBlockChan:
-			handler.SafelyHandleMessage(c.ctx, c.processBlock, block)
-		}
-	}
-}
-
-func (c *ChainService) processBlock(message proto.Message) {
-	block := message.(*pb.BeaconBlock)
-	beaconState, err := c.beaconDB.State(c.ctx)
-	if err != nil {
-		log.Errorf("Unable to retrieve beacon state %v", err)
-		return
-	}
-	if block.Slot > beaconState.Slot {
-		computedState, err := c.ReceiveBlock(block, beaconState)
-		if err != nil {
-			log.Errorf("Could not process received block: %v", err)
-			return
-		}
-		if err := c.ApplyForkChoiceRule(block, computedState); err != nil {
-			log.Errorf("Could not update chain head: %v", err)
-			return
-		}
-	}
-}
-
-// ApplyForkChoiceRule determines the current beacon chain head using LMD GHOST as a block-vote
-// weighted function to select a canonical head in Ethereum Serenity.
-func (c *ChainService) ApplyForkChoiceRule(block *pb.BeaconBlock, computedState *pb.BeaconState) error {
-	h, err := hashutil.HashBeaconBlock(block)
-	if err != nil {
-		return fmt.Errorf("could not tree hash incoming block: %v", err)
-	}
-	// TODO(#1307): Use LMD GHOST as the fork-choice rule for Ethereum Serenity.
-	// TODO(#674): Handle chain reorgs.
-	if err := c.beaconDB.UpdateChainHead(block, computedState); err != nil {
-		return fmt.Errorf("failed to update chain: %v", err)
-	}
-	log.WithField("blockRoot", fmt.Sprintf("0x%x", h)).Info("Chain head block and state updated")
-	// We fire events that notify listeners of a new block in
-	// the case of a state transition. This is useful for the beacon node's gRPC
-	// server to stream these events to beacon clients.
-	// When the transition is a cycle transition, we stream the state containing the new validator
-	// assignments to clients.
-	if helpers.IsEpochStart(block.Slot) {
-		if c.canonicalStateFeed.Send(computedState) == 0 {
-			log.Error("Sent canonical state to no subscribers")
-		}
-	}
-
-	if err := c.saveFinalizedState(computedState); err != nil {
-		log.Errorf("Could not save new finalized state: %v", err)
-	}
-	if c.canonicalBlockFeed.Send(&pb.BeaconBlockAnnounce{
-		Hash:       h[:],
-		SlotNumber: block.Slot,
-	}) == 0 {
-		log.Error("Sent canonical block to no subscribers")
-	}
-	return nil
-}
-
-// ReceiveBlock is a function that defines the operations that are preformed on
-// any block that is received from p2p layer or rpc. It checks the block to see
-// if it passes the pre-processing conditions, if it does then the per slot
-// state transition function is carried out on the block.
-// spec:
-//  def process_block(block):
-//      if not block_pre_processing_conditions(block):
-//          return nil, error
-//
-//  	# process skipped slots
-//
-// 		while (state.slot < block.slot - 1):
-//      	state = slot_state_transition(state, block=None)
-//
-//		# process slot with block
-//		state = slot_state_transition(state, block)
-//
-//		# check state root
-//      if block.state_root == hash(state):
-//			return state, error
-//		else:
-//			return nil, error  # or throw or whatever
-//
-func (c *ChainService) ReceiveBlock(block *pb.BeaconBlock, beaconState *pb.BeaconState) (*pb.BeaconState, error) {
-	blockRoot, err := hashutil.HashBeaconBlock(block)
-	if err != nil {
-		return nil, fmt.Errorf("could not tree hash incoming block: %v", err)
-	}
-
-	if block.Slot == params.BeaconConfig().GenesisSlot {
-		return nil, fmt.Errorf("cannot process a genesis block: received block with slot %d",
-			block.Slot-params.BeaconConfig().GenesisSlot)
-	}
-
-	// Save blocks with higher slot numbers in cache.
-	if err := c.isBlockReadyForProcessing(block, beaconState); err != nil {
-		return nil, fmt.Errorf("block with root %#x is not ready for processing: %v", blockRoot, err)
-	}
-
-	// Retrieve the last processed beacon block's hash root.
-	headRoot, err := c.ChainHeadRoot()
-	if err != nil {
-		return nil, fmt.Errorf("could not retrieve chain head root: %v", err)
-	}
-
-	log.WithField("slotNumber", block.Slot-params.BeaconConfig().GenesisSlot).Info(
-		"Executing state transition")
-
-	// Check for skipped slots.
-	numSkippedSlots := 0
-	for beaconState.Slot < block.Slot-1 {
-		beaconState, err = c.runStateTransition(headRoot, nil, beaconState)
-		if err != nil {
-			return nil, fmt.Errorf("could not execute state transition without block %v", err)
-		}
-		numSkippedSlots++
-	}
-	if numSkippedSlots > 0 {
-		log.Warnf("Processed %d skipped slots", numSkippedSlots)
-	}
-
-	beaconState, err = c.runStateTransition(headRoot, block, beaconState)
-	if err != nil {
-		return nil, fmt.Errorf("could not execute state transition with block %v", err)
-	}
-
-	// if there exists a block for the slot being processed.
-	if err := c.beaconDB.SaveBlock(block); err != nil {
-		return nil, fmt.Errorf("failed to save block: %v", err)
-	}
-
-	// Forward processed block to operation pool to remove individual operation from DB.
-	if c.opsPoolService.IncomingProcessedBlockFeed().Send(block) == 0 {
-		log.Error("Sent processed block to no subscribers")
-	}
-
-	// Remove pending deposits from the deposit queue.
-	for _, dep := range block.Body.Deposits {
-		c.beaconDB.RemovePendingDeposit(c.ctx, dep)
-	}
-
-	log.WithField("hash", fmt.Sprintf("%#x", blockRoot)).Debug("Processed beacon block")
-	return beaconState, nil
-}
-
-func (c *ChainService) runStateTransition(headRoot [32]byte, block *pb.BeaconBlock,
-	beaconState *pb.BeaconState) (*pb.BeaconState, error) {
-	beaconState, err := state.ExecuteStateTransition(
-		c.ctx,
-		beaconState,
-		block,
-		headRoot,
-		true, /* sig verify */
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not execute state transition %v", err)
-	}
-	log.WithField(
-		"slotsSinceGenesis", beaconState.Slot-params.BeaconConfig().GenesisSlot,
-	).Info("Slot transition successfully processed")
-
-	if block != nil {
-		log.WithField(
-			"slotsSinceGenesis", beaconState.Slot-params.BeaconConfig().GenesisSlot,
-		).Info("Block transition successfully processed")
-	}
-
-	if helpers.IsEpochEnd(beaconState.Slot) {
-		// Save activated validators of this epoch to public key -> index DB.
-		if err := c.saveValidatorIdx(beaconState); err != nil {
-			return nil, fmt.Errorf("could not save validator index: %v", err)
-		}
-		// Delete exited validators of this epoch to public key -> index DB.
-		if err := c.deleteValidatorIdx(beaconState); err != nil {
-			return nil, fmt.Errorf("could not delete validator index: %v", err)
-		}
-		log.WithField(
-			"SlotsSinceGenesis", beaconState.Slot-params.BeaconConfig().GenesisSlot,
-		).Info("Epoch transition successfully processed")
-	}
-	return beaconState, nil
-}
-
-func (c *ChainService) isBlockReadyForProcessing(block *pb.BeaconBlock, beaconState *pb.BeaconState) error {
-	var powBlockFetcher func(ctx context.Context, hash common.Hash) (*gethTypes.Block, error)
-	if c.enablePOWChain {
-		powBlockFetcher = c.web3Service.Client().BlockByHash
-	}
-	if err := b.IsValidBlock(c.ctx, beaconState, block, c.enablePOWChain,
-		c.beaconDB.HasBlock, powBlockFetcher, c.genesisTime); err != nil {
-		return fmt.Errorf("block does not fulfill pre-processing conditions %v", err)
-	}
-	return nil
-}
-
-func (c *ChainService) saveFinalizedState(beaconState *pb.BeaconState) error {
-	// check if the finalized epoch has changed, if it
-	// has we save the finalized state.
-	if c.finalizedEpoch != beaconState.FinalizedEpoch {
-		c.finalizedEpoch = beaconState.FinalizedEpoch
-		return c.beaconDB.SaveFinalizedState(beaconState)
-	}
-	return nil
-}
-
-// saveValidatorIdx saves the validators public key to index mapping in DB, these
-// validators were activated from current epoch. After it saves, current epoch key
-// is deleted from ActivatedValidators mapping.
-func (c *ChainService) saveValidatorIdx(state *pb.BeaconState) error {
-	for _, idx := range validators.ActivatedValidators[helpers.CurrentEpoch(state)] {
-		pubKey := state.ValidatorRegistry[idx].Pubkey
-		if err := c.beaconDB.SaveValidatorIndex(pubKey, int(idx)); err != nil {
-			return fmt.Errorf("could not save validator index: %v", err)
-		}
-	}
-	delete(validators.ActivatedValidators, helpers.CurrentEpoch(state))
-	return nil
-}
-
-// deleteValidatorIdx deletes the validators public key to index mapping in DB, the
-// validators were exited from current epoch. After it deletes, current epoch key
-// is deleted from ExitedValidators mapping.
-func (c *ChainService) deleteValidatorIdx(state *pb.BeaconState) error {
-	for _, idx := range validators.ExitedValidators[helpers.CurrentEpoch(state)] {
-		pubKey := state.ValidatorRegistry[idx].Pubkey
-		if err := c.beaconDB.DeleteValidatorIndex(pubKey); err != nil {
-			return fmt.Errorf("could not delete validator index: %v", err)
-		}
-	}
-	delete(validators.ExitedValidators, helpers.CurrentEpoch(state))
-	return nil
-}
-
-// attestationTargets retrieves the list of attestation targets since last finalized epoch,
-// each attestation target consists of validator index and its attestation target (i.e. the block
-// which the validator attested to)
-func (c *ChainService) attestationTargets(state *pb.BeaconState) ([]*attestationTarget, error) {
-	indices := helpers.ActiveValidatorIndices(state.ValidatorRegistry, state.FinalizedEpoch)
-	attestationTargets := make([]*attestationTarget, len(indices))
-	for i, index := range indices {
-		block, err := c.attsService.LatestAttestationTarget(c.ctx, index)
-		if err != nil {
-			return nil, fmt.Errorf("could not retrieve attestation target: %v", err)
-		}
-		attestationTargets[i] = &attestationTarget{
-			validatorIndex: index,
-			block:          block,
-		}
-	}
-	return attestationTargets, nil
-}
-
-// blockChildren returns the child blocks of the given block.
-// ex:
-//       /- C - E
-// A - B - D - F
-//       \- G
-// Input: B. Output: [C, D, G]
-//
-// Spec pseudocode definition:
-//	get_children(store: Store, block: BeaconBlock) -> List[BeaconBlock]
-//		returns the child blocks of the given block.
-func (c *ChainService) blockChildren(block *pb.BeaconBlock, state *pb.BeaconState) ([]*pb.BeaconBlock, error) {
-	var children []*pb.BeaconBlock
-
-	currentRoot, err := hashutil.HashBeaconBlock(block)
-	if err != nil {
-		return nil, fmt.Errorf("could not tree hash incoming block: %v", err)
-	}
-	startSlot := block.Slot + 1
-	currentSlot := state.Slot
-	for i := startSlot; i <= currentSlot; i++ {
-		block, err := c.beaconDB.BlockBySlot(i)
-		if err != nil {
-			return nil, fmt.Errorf("could not get block by slot: %v", err)
-		}
-		// Continue if there's a skip block.
-		if block == nil {
-			continue
-		}
-
-		parentRoot := bytesutil.ToBytes32(block.ParentRootHash32)
-		if currentRoot == parentRoot {
-			children = append(children, block)
-		}
-	}
-	return children, nil
-}
-
-// lmdGhost applies the Latest Message Driven, Greediest Heaviest Observed Sub-Tree
-// fork-choice rule defined in the Ethereum Serenity specification for the beacon chain.
-//
-// Spec pseudocode definition:
-//	def lmd_ghost(store: Store, start_state: BeaconState, start_block: BeaconBlock) -> BeaconBlock:
-//    """
-//    Execute the LMD-GHOST algorithm to find the head ``BeaconBlock``.
-//    """
-//    validators = start_state.validator_registry
-//    active_validator_indices = get_active_validator_indices(validators, slot_to_epoch(start_state.slot))
-//    attestation_targets = [
-//        (validator_index, get_latest_attestation_target(store, validator_index))
-//        for validator_index in active_validator_indices
-//    ]
-//
-//    def get_vote_count(block: BeaconBlock) -> int:
-//        return sum(
-//            get_effective_balance(start_state.validator_balances[validator_index]) // FORK_CHOICE_BALANCE_INCREMENT
-//            for validator_index, target in attestation_targets
-//            if get_ancestor(store, target, block.slot) == block
-//        )
-//
-//    head = start_block
-//    while 1:
-//        children = get_children(store, head)
-//        if len(children) == 0:
-//            return head
-//        head = max(children, key=get_vote_count)
-func (c *ChainService) lmdGhost(
-	block *pb.BeaconBlock,
-	state *pb.BeaconState,
-	voteTargets map[uint64]*pb.BeaconBlock,
-) (*pb.BeaconBlock, error) {
-	head := block
-	for {
-		children, err := c.blockChildren(head, state)
-		if err != nil {
-			return nil, fmt.Errorf("could not fetch block children: %v", err)
-		}
-		if len(children) == 0 {
-			return head, nil
-		}
-		maxChild := children[0]
-
-		maxChildVotes, err := VoteCount(maxChild, state, voteTargets, c.beaconDB)
-		if err != nil {
-			return nil, fmt.Errorf("unable to determine vote count for block: %v", err)
-		}
-		for i := 0; i < len(children); i++ {
-			candidateChildVotes, err := VoteCount(children[i], state, voteTargets, c.beaconDB)
-			if err != nil {
-				return nil, fmt.Errorf("unable to determine vote count for block: %v", err)
-			}
-			if candidateChildVotes > maxChildVotes {
-				maxChild = children[i]
-			}
-		}
-		head = maxChild
-	}
 }
